Refactor authentication forms to use new useFormSubmission hook
