+++ Client/src/app/auth/sign-up/page.tsx
@@ -12,7 +12,7 @@ import AuthFormWrapper from '../components/AuthFormWrapper';
 import FormInput from '../../../components/FormInput';
 import PasswordValidation from '../components/PasswordValidation';
 
-import { useAuthForm } from '@/hooks/useAuthForm';
+import { useFormSubmission } from '@/hooks/useFormSubmission';
 import { useValidatedFormData } from '@/hooks/useValidatedFormData';
 import { minLengthRule, requiredFieldRule, validEmailRule } from '@/utils/shared/validators';
 
@@ -36,7 +36,7 @@ export default function SignUp() {
 		},
 	});
 
-	const { loading, handleSubmit, toast } = useAuthForm({
+	const { loading, handleSubmit, toast } = useFormSubmission({
 		onSubmit: async () => {
 			// 1) Basic client checks
 			const hasError = validateAll();
@@ -55,7 +55,7 @@ export default function SignUp() {
 			});
 
 			if (res.data.success) {
-				// Partial success?
+				// Partial success
 				if (res.data.emailFail) {
 					toast.showToast({
 						message:
@@ -63,19 +63,13 @@ export default function SignUp() {
 							'Account created, Email sending is disabled in development. Contact admin.',
 						variant: 'warning',
 					});
-					// We have userId => poll by userId
 					return router.push(`/auth/account-created?userId=${res.data.userId}`);
 				}
 
-				// Full success => poll by token or userId
 				if (res.data.token) {
-					// If the server returned a token, let's poll by token:
 					router.push(`/auth/account-created?token=${res.data.token}`);
 				} else {
-					// Or if no token was returned, fallback to userId or a success message
 					toast.showToast({ message: res.data.message, variant: 'success' });
-					// Possibly poll by userId if you want
-					// router.push(`/auth/account-created?userId=someUserId`);
 				}
 			} else {
 				throw new Error(res.data.message || 'Unknown server error');
